requirements.txt
Django==3.2
djangorestframework==3.12.4
django-environ==0.4.5
flake8==3.9.1
black==21.4b2
psycopg2-binary==2.8.6
channels==3.0.4
pytest-django==4.1.0
pytest-cov==2.11.1
pillow==8.3.2